Add exists check for sqlite3_analyzer

+++ RapidBuild64/002-utils/03c-clean_sqlite.sh
@@ -4,6 +4,6 @@ source ../spec.sh
 source ./module_spec.sh
 
 # clean up sqlite, if we don't have tcl
-if [ "${INCTCL}" != "Y" ]; then
+if [ -f ${DIR}/usr/bin/sqlite3_analyzer ] && [ "${INCTCL}" != "Y" ]; then
   rm ${DIR}/usr/bin/sqlite3_analyzer
 fi
